# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aadex HMS API",
        "description": "Multi-tenant chain hotel management API.\n\n**This page currently documents the Authentication API only.** Roles\n(`super_admin`, `hotel_chain_admin`, `hotel_admin`, `staff`) and the\nper-module rights taxonomy are seeded and enforced in the database — see\nthe README for the model — but the Super Admin, Chain and Property\nendpoints belong to later phases and are not published here yet.\n\n**To authorise:** call `POST /auth/login`, copy `token` from the response,\nthen click **Authorize** above and paste it as the bearer token.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new hotel account (awaits Super Admin approval)",
                "description": "Public self-signup. Creates the hotel and its Hotel Admin user, both with status `pending` — no token is issued, and `POST /auth/login` answers 403 until a Super Admin approves the account. (The approval endpoint lives in the Super Admin module, which is not published on this page yet.)",
                "operationId": "1ad4e4447f4ef3eb632ed2691c00c0e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "hotel_name",
                                    "admin_name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "hotel_name": {
                                        "type": "string"
                                    },
                                    "admin_name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "gst_tax_id": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "INR",
                                        "nullable": true
                                    },
                                    "timezone": {
                                        "type": "string",
                                        "example": "Asia/Kolkata",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registered, pending approval"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "429": {
                        "description": "Too many signup attempts"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log in and receive a Sanctum API token",
                "description": "Returns `{ token, expires_at, user }`. `expires_at` is when the token dies (from `sanctum.expiration`, default 7 days) — clients should re-login before then. `device_name` labels the session in `GET /auth/sessions`.",
                "operationId": "e5e32c5c121d50e864abf0f704b71593",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authenticated"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Account inactive or pending approval"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current API token",
                "operationId": "1536ec034ecbb551483aeacfd4eea588",
                "responses": {
                    "200": {
                        "description": "Logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get the authenticated user",
                "operationId": "bf21b27a7e826768548b5b14ee2d44f2",
                "responses": {
                    "200": {
                        "description": "Current user"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change the authenticated user's own password",
                "description": "Verifies the current password, then revokes every *other* token for this user so other devices are signed out. The token used for this call stays valid.",
                "operationId": "3f30e842d7e1e0a796e94a61391bcadf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed"
                    },
                    "422": {
                        "description": "Current password incorrect, or new password invalid"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/forgot-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Request a password reset OTP by email",
                "operationId": "74a9d56dc98a50ca3d03f4781373895f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OTP sent if the account exists"
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password using an OTP",
                "operationId": "13fa90fc3ea9d8db4992c94b3c12da80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "otp": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password reset"
                    },
                    "422": {
                        "description": "Invalid or expired OTP"
                    }
                }
            }
        },
        "/auth/sessions": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "List the current user's active sessions (devices)",
                "description": "One row per live Sanctum token, with the IP and user agent captured at login. Expired tokens are filtered out even before `sanctum:prune-expired` deletes them.",
                "operationId": "605384dcb3c1631175f5068e8ee1cb7f",
                "responses": {
                    "200": {
                        "description": "Active sessions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/sessions/{session}": {
            "delete": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke one of the current user's sessions",
                "description": "Signs that device out. Revoking the session you are calling from is allowed and is equivalent to logging out — the response says so via `revoked_current`.",
                "operationId": "896e0ead9ed8f34a5dfd4af0ab39b6a6",
                "parameters": [
                    {
                        "name": "session",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Session revoked"
                    },
                    "404": {
                        "description": "No such session for this user"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/auth/logout-all": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign out of every device",
                "description": "Revokes every token for the current user, including the one making the call — the client must log in again afterwards.",
                "operationId": "e388cc825f3fec49f90c6cf613ec88d3",
                "responses": {
                    "200": {
                        "description": "All sessions revoked"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Sanctum personal access token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}